move scraper file to lib folder
